// src/app/shared/constants/rol.enum.ts
export enum ERol
{
  ADMIN="ADMIN",
  VALIDA_ALUMNO = "VALIDA_ALUMNO",
  VALIDA_CURSO = "VALIDA_CURSO",
  VALIDA_PLANTEL = "VALIDA_PLANTEL",
  VALIDA_DOCENTE = "VALIDA_DOCENTE",
  DOCENTE = "DOCENTE",
  ALUMNO = "ALUMNO",
  CONTROL_ESCOLAR = "CONTROL_ESCOLAR",
  COORDINADOR_ACADEMICO = "COORDINADOR_ACADEMICO",
  ADMIN_FINANZAS="ADMIN_FINANZAS",
  PLANTEL="PLANTEL",
}